COTECH-787 extract filterVideoBids() to the shared folder
