+++ src/platforms/shared/bidders/filter-video-bids.ts
@@ -1,0 +1,23 @@
+import { context } from '@wikia/ad-engine';
+
+export function filterVideoBids(bidderContext) {
+	const hasFeaturedVideo = context.get('custom.hasFeaturedVideo');
+	const bidConfigSlotNames = Object.keys(bidderContext.slots);
+	const hasFeaturedBidConfig = bidderContext.slots && bidConfigSlotNames.includes('featured');
+	const hasIncontentPlayerBidConfig =
+		bidderContext.slots && bidConfigSlotNames.includes('incontent_player');
+	const newVideoSlotsConfig = {
+		...bidderContext.slots,
+	};
+
+	if (hasFeaturedVideo && hasIncontentPlayerBidConfig) {
+		delete newVideoSlotsConfig['incontent_player'];
+	} else if (!hasFeaturedVideo && hasFeaturedBidConfig) {
+		delete newVideoSlotsConfig['featured'];
+	}
+
+	return {
+		...bidderContext,
+		slots: newVideoSlotsConfig,
+	};
+}
