+++ src/platforms/shared/index.ts
@@ -1,4 +1,5 @@
 export * from './bidders/bidders-state.setup';
+export * from './bidders/filter-video-bids';
 export * from './bidders/wikia-adapter';
 export * from './bootstrap';
 export * from './context/targeting/targeting-strategies/models/open-rtb2';
